VMSVGA3d: Corresponding win build fixes.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@57083 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Graphics/DevVGA-SVGA3d-shared.cpp
@@ -443,7 +443,7 @@ const char *vmsvgaLookupEnum(int32_t iValue, PCVMSVGAINFOENUMMAP pEnumMap)
      * Binary search
      */
     uint32_t iStart = 0;
-    uint32_t iEnd   = pEnumMap->cValues;
+    uint32_t iEnd   = (uint32_t)pEnumMap->cValues;
     for (;;)
     {
         uint32_t i = iStart + (iEnd - iStart) / 2;
